+++ django_tables2/__init__.py
@@ -8,4 +8,4 @@ from .utils import A
 from .views import SingleTableMixin, SingleTableView, MultiTableMixin
 
 
-__version__ = '1.2.3'
+__version__ = '1.2.5'