+++ django_tables2/columns/booleancolumn.py
@@ -35,12 +35,13 @@ class BooleanColumn(Column):
         super(BooleanColumn, self).__init__(**kwargs)
 
     def render(self, value, record, bound_column):
-        # if record is a model, we need to check if it has choices defined.
-        # If that's the case, we need to inverse lookup the value to convert to
-        # a boolean.
+        # If record is a model, we need to check if it has choices defined.
         if hasattr(record, '_meta'):
             field = bound_column.accessor.get_field(record)
-            if hasattr(field, 'choices') and field.choices is not None:
+
+            # If that's the case, we need to inverse lookup the value to convert
+            # to a boolean we can use.
+            if hasattr(field, 'choices') and field.choices is not None and len(field.choices) > 0:
                 value = next(val for val, name in field.choices if name == value)
 
         value = bool(value)